Add pagination for flickr.photos.getContactsPhotos
(also added as comment in the code):

The api docs don't mention that pagination for this method is supported.
However, looking at the json response from the api explorer it quite
clearly is.  I have tested and it works fine.

It's likely Flickr just need to update their docs, though this may
change in the future.